Examples

+++ examples/SI47XX_08_RDS/SI47XX_02_RDS_TFT_TOUCH_SHIELD/SI47XX_02_RDS_TFT_TOUCH_SHIELD.ino
@@ -497,7 +497,7 @@ void showFrequency()
 }
 
 // Will be used by seekStationProgress 
-void showFrequencySeek(uint8_t freq) {
+void showFrequencySeek(uint16_t freq) {
   currentFrequency = freq;
   showFrequency();
 }
@@ -859,6 +859,7 @@ void loop(void)
     // bSeekUp.drawButton(true);
     if (currentMode == FM) {
       si4735.seekStationProgress(showFrequencySeek,1);
+      // si4735.seekStationUp();
       delay(15);
       currentFrequency = si4735.getFrequency();
     }
@@ -872,6 +873,7 @@ void loop(void)
     // bSeekUp.drawButton(true);
     if (currentMode == FM) {
       si4735.seekStationProgress(showFrequencySeek, 0);
+      // si4735.seekStationDown();
       delay(15);
       currentFrequency = si4735.getFrequency();
     }
